fix: verify and apply audit findings across shared/bot modules
- Lazy-init webPrisma client and auth-node (no eager Prisma at import time)

- Tighten snowflake validation with BigInt range check

- Add JSON roundtrip safety to normalizeDiscordLogForDatabase

- Move audit-log to src/shared (shared code no longer imports from lib/)

- Simplify permissions writeCached post-purge (no delete/recreate cycle)

- Reject NaN/Infinity/exact-expiry in ws-connect-token parseWsConnectToken

- Move DISCORD_BOT_INVITE_PERMISSIONS to src/shared (no discord.js dep)

- Remove double response on invalid days in /downloads cleanup

Made-with: Cursor
